requirements.txt
opencv-python~=4.8.1.78
numpy~=1.26.2
pillow~=10.0.1
pandas~=2.1.4
matplotlib~=3.8.0